move serialize_form and clean_up_url_parameters to UrlSubmitForm
This is the only file using these function, so make them local.
